refactor(transcripts.tsx): improve button label for adding new transcript based on task type
The button label for adding a new transcript has been updated to be more descriptive based on the task type. If the task type is 'VOICE TO TEXT', the label will be translated using the translation function 't' as 'add-new-transcript'. Otherwise, the label will be 'Add new Sentence'. This change improves the clarity and usability of the button.