feat(create-admin.js): add script to create an admin user in the database
The `create-admin.js` script is added to the project. This script uses the `PrismaClient` and `bcrypt` libraries to create an admin user in the database. The admin user has the following details:
- Email: info@akilli.ai
- Password: Default123@
- Mobile Number: 0022556525141
- Country: Benin
- Role: ADMIN
- Name: Manager

The script hashes the password using bcrypt before storing it in the database. If any error occurs during the creation of the admin user, it will be logged to the console.